Added restore, added safeties to improper usage

+++ freezerestore/freeze.py
@@ -1,3 +1,8 @@
 import shutil
+import os.path
+from os import path
+
 def freeze(filetitle):
+    if not path.exists(filetitle):
+        return
     shutil.copy(filetitle,filetitle + ".temp")
